requirements.txt
Flask
gunicorn
requests
beautifulsoup4
nltk
newspaper3k
spacy
wordcloud
lxml
pillow
